html/requirements.txt
Django==2.2.14
django-autoslug==1.9.7
django-crispy-forms==1.9.2
pytz==2020.1
sqlparse==0.3.1
mysqlclient
